Move system_events models
