--- botx/bot/models/commands/system_events/added_to_chat.py
@@ -1,18 +1,0 @@
-from dataclasses import dataclass
-from typing import List
-from uuid import UUID
-
-from botx.bot.models.commands.base import BotCommandBase
-from botx.bot.models.commands.chat import Chat
-
-
-@dataclass
-class AddedToChatEvent(BotCommandBase):
-    """Event `system:added_to_chat`.
-
-    Attributes:
-        huids: List of added to chat user huids.
-    """
-
-    huids: List[UUID]
-    chat: Chat
