+++ botx/models/system_events/added_to_chat.py
@@ -1,3 +1,4 @@
+from dataclasses import dataclass
 from typing import Any, Dict, List
 from uuid import UUID
 
@@ -5,10 +6,8 @@ from pydantic import Field
 
 from botx.bot.api.commands.base import BotAPIBaseCommand, BotAPIChatContext
 from botx.bot.api.enums import BotAPICommandTypes
+from botx.bot.models.commands.base import BotCommandBase
 from botx.bot.models.commands.chat import Chat
-from botx.bot.models.commands.system_events.deleted_from_chat import (
-    DeletedFromChatEvent,
-)
 from botx.shared_models.api_base import VerifiedPayloadBaseModel
 from botx.shared_models.chat_types import convert_chat_type_to_domain
 
@@ -18,25 +17,38 @@ except ImportError:
     from typing_extensions import Literal  # type: ignore  # noqa: WPS440
 
 
-class BotAPIDeletedFromChatData(VerifiedPayloadBaseModel):
-    deleted_members: List[UUID]
+@dataclass
+class AddedToChatEvent(BotCommandBase):
+    """Event `system:added_to_chat`.
+
+    Attributes:
+        huids: List of added to chat user huids.
+    """
+
+    huids: List[UUID]
+    chat: Chat
 
 
-class BotAPIDeletedFromChatPayload(VerifiedPayloadBaseModel):
-    body: Literal["system:deleted_from_chat"] = "system:deleted_from_chat"
+class BotAPIAddedToChatData(VerifiedPayloadBaseModel):
+    added_members: List[UUID]
+
+
+class BotAPIAddedToChatPayload(VerifiedPayloadBaseModel):
+    body: Literal["system:added_to_chat"] = "system:added_to_chat"
     command_type: Literal[BotAPICommandTypes.SYSTEM]
-    data: BotAPIDeletedFromChatData
+    data: BotAPIAddedToChatData
+    metadata: Dict[str, Any]
 
 
-class BotAPIDeletedFromChat(BotAPIBaseCommand):
-    payload: BotAPIDeletedFromChatPayload = Field(..., alias="command")
+class BotAPIAddedToChat(BotAPIBaseCommand):
+    payload: BotAPIAddedToChatPayload = Field(..., alias="command")
     sender: BotAPIChatContext = Field(..., alias="from")
 
-    def to_domain(self, raw_command: Dict[str, Any]) -> DeletedFromChatEvent:
-        return DeletedFromChatEvent(
+    def to_domain(self, raw_command: Dict[str, Any]) -> AddedToChatEvent:
+        return AddedToChatEvent(
             bot_id=self.bot_id,
             raw_command=raw_command,
-            huids=self.payload.data.deleted_members,
+            huids=self.payload.data.added_members,
             chat=Chat(
                 id=self.sender.group_chat_id,
                 type=convert_chat_type_to_domain(self.sender.chat_type),
