+++ botx/models/system_events/deleted_from_chat.py
@@ -1,3 +1,4 @@
+from dataclasses import dataclass
 from typing import Any, Dict, List
 from uuid import UUID
 
@@ -5,8 +6,8 @@ from pydantic import Field
 
 from botx.bot.api.commands.base import BotAPIBaseCommand, BotAPIChatContext
 from botx.bot.api.enums import BotAPICommandTypes
+from botx.bot.models.commands.base import BotCommandBase
 from botx.bot.models.commands.chat import Chat
-from botx.bot.models.commands.system_events.added_to_chat import AddedToChatEvent
 from botx.shared_models.api_base import VerifiedPayloadBaseModel
 from botx.shared_models.chat_types import convert_chat_type_to_domain
 
@@ -16,26 +17,38 @@ except ImportError:
     from typing_extensions import Literal  # type: ignore  # noqa: WPS440
 
 
-class BotAPIAddedToChatData(VerifiedPayloadBaseModel):
-    added_members: List[UUID]
+@dataclass
+class DeletedFromChatEvent(BotCommandBase):
+    """Event `system:deleted_from_chat`.
+
+    Attributes:
+        huids: List of deleted from chat user huids.
+        chat_id: Chat where the user was deleted from.
+    """
+
+    huids: List[UUID]
+    chat: Chat
 
 
-class BotAPIAddedToChatPayload(VerifiedPayloadBaseModel):
-    body: Literal["system:added_to_chat"] = "system:added_to_chat"
+class BotAPIDeletedFromChatData(VerifiedPayloadBaseModel):
+    deleted_members: List[UUID]
+
+
+class BotAPIDeletedFromChatPayload(VerifiedPayloadBaseModel):
+    body: Literal["system:deleted_from_chat"] = "system:deleted_from_chat"
     command_type: Literal[BotAPICommandTypes.SYSTEM]
-    data: BotAPIAddedToChatData
-    metadata: Dict[str, Any]
+    data: BotAPIDeletedFromChatData
 
 
-class BotAPIAddedToChat(BotAPIBaseCommand):
-    payload: BotAPIAddedToChatPayload = Field(..., alias="command")
+class BotAPIDeletedFromChat(BotAPIBaseCommand):
+    payload: BotAPIDeletedFromChatPayload = Field(..., alias="command")
     sender: BotAPIChatContext = Field(..., alias="from")
 
-    def to_domain(self, raw_command: Dict[str, Any]) -> AddedToChatEvent:
-        return AddedToChatEvent(
+    def to_domain(self, raw_command: Dict[str, Any]) -> DeletedFromChatEvent:
+        return DeletedFromChatEvent(
             bot_id=self.bot_id,
             raw_command=raw_command,
-            huids=self.payload.data.added_members,
+            huids=self.payload.data.deleted_members,
             chat=Chat(
                 id=self.sender.group_chat_id,
                 type=convert_chat_type_to_domain(self.sender.chat_type),
